% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATCH_FDIex.R
\docType{data}
\name{CATCH_FDIex}
\alias{CATCH_FDIex}
\title{CATCH_FDIex : example of Catch table FDI}
\format{
An object of class \code{data.frame} with 2540 rows and 23 columns.
}
\usage{
CATCH_FDIex
}
\description{
CATCH_FDIex : example of Catch table FDI
}
\author{
Isabella Bitetto \email{bitetto@coispa.it}
}
\keyword{Catch}
\keyword{FDI}
