% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATCH_MEDBS.R
\name{CATCH_MEDBS}
\alias{CATCH_MEDBS}
\title{Catch at age (CATCH) table - MED & BS data call}
\usage{
CATCH_MEDBS(datacs, datacl, datace, verbose = FALSE)
}
\arguments{
\item{datacs}{Detailed data in RCG CS format}

\item{datacl}{Landings aggregated data in RCG CL format}

\item{datace}{Effort aggregated data in RDB CE format (https://www.ices.dk/data/Documents/RDB/RDB\%20Exchange\%20Format.pdf)}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
CATCH table
}
\description{
Catch at age (CATCH) table - MED & BS data call
}
\examples{
CATCH_MEDBS(RDBprocessing::data_ex,RDBprocessing::data_exampleCL,RDBprocessing::ce_example)
}
