% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CT.R
\docType{data}
\name{CT}
\alias{CT}
\title{CT communication table GFCM-MEDBS-FDI}
\format{
An object of class \code{data.frame} with 134 rows and 7 columns.
}
\usage{
CT
}
\description{
CT communication table GFCM-MEDBS-FDI
}
\author{
Isabella Bitetto \email{bitetto@coispa.it}
}
\keyword{GFCM_fleet_segments}
