% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery.R
\docType{data}
\name{fishery}
\alias{fishery}
\title{communicationTable_for_fishery}
\format{
An object of class \code{data.frame} with 13 rows and 2 columns.
}
\usage{
fishery
}
\description{
communicationTable_for_fishery
}
\author{
Isabella Bitetto \email{bitetto@coispa.it}
}
\keyword{communicationTable_for_fishery}
