% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FT_GEAR.R
\docType{data}
\name{FT_GEAR}
\alias{FT_GEAR}
\title{communication table fishing technique- gear based on Italian FDI data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19 rows and 2 columns.
}
\usage{
FT_GEAR
}
\description{
communication table fishing technique- gear based on Italian FDI data
}
\author{
Isabella Bitetto \email{bitetto@coispa.it}
}
\keyword{FDI}
\keyword{fishing}
\keyword{gear}
\keyword{technique}
