% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_MEDBS.R
\name{ML_MEDBS}
\alias{ML_MEDBS}
\title{Maturity at length (ML) table - MED & BS data call}
\usage{
ML_MEDBS(data, imm = c("1", "2", "2a"))
}
\arguments{
\item{data}{Detailed data in RCG CS format}

\item{imm}{maturity stages to be considered as immature}
}
\value{
ML table
}
\description{
Maturity at length (ML) table - MED & BS data call
}
\examples{
ML_MEDBS(RDBprocessing::data_ex)
}
