% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msr.R
\docType{data}
\name{msr}
\alias{msr}
\title{Mesh size range codification MED &  BS data call}
\format{
An object of class \code{data.frame} with 8 rows and 2 columns.
}
\usage{
msr
}
\description{
Mesh size range codification MED &  BS data call
}
\author{
Isabella Bitetto \email{bitetto@coispa.it}
}
\keyword{mesh_size_code}
