% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCGtoCOST_CL.R
\name{RCGtoCOST_CL}
\alias{RCGtoCOST_CL}
\title{Function converting RCG CL in COST CL object}
\usage{
RCGtoCOST_CL(data, verbose = FALSE)
}
\arguments{
\item{data}{Landing data in RCG CL format}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
COST CL object
}
\description{
Function converting RCG CL in COST CL object
}
\examples{
RCGtoCOST_CL(RDBprocessing::data_exampleCL)
}
