% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabII2_GFCM.R
\name{TabII2_GFCM}
\alias{TabII2_GFCM}
\title{Table II.2  - GFCM DCRF data call}
\usage{
TabII2_GFCM(CatchFDI, verbose = FALSE)
}
\arguments{
\item{CatchFDI}{Catch table FDI}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
Table II.2 GFCM DCRF table
}
\description{
Table II.2  - GFCM DCRF data call
}
\examples{
TabII2_GFCM(RDBprocessing::CATCH_FDIex)
}
