% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRL_MEDBS.R
\name{SRL_MEDBS}
\alias{SRL_MEDBS}
\title{Sex ratio at length (SRL) table - MED & BS data call}
\usage{
SRL_MEDBS(data)
}
\arguments{
\item{data}{Detailed data in RCG CS format}
}
\value{
SRL table
}
\description{
Sex ratio at length (SRL) table - MED & BS data call
}
\examples{
SRL_MEDBS(RDBprocessing::data_ex)
}
